refactor : 투표 DTO 역할별 분리
