+++ ddv/src/main/java/community/ddv/domain/vote/controller/VoteController.java
@@ -2,9 +2,9 @@ package community.ddv.domain.vote.controller;
 
 import community.ddv.domain.user.entity.User;
 import community.ddv.domain.user.service.UserService;
-import community.ddv.domain.vote.dto.VoteDTO.VoteResultDTO;
+import community.ddv.domain.vote.dto.VoteResultDTO;
 import community.ddv.domain.vote.dto.VoteOptionsDTO;
-import community.ddv.domain.vote.dto.VoteParticipationDTO.VoteParticipationRequestDto;
+import community.ddv.domain.vote.dto.VoteParticipationDTO;
 import community.ddv.domain.vote.service.VoteService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
@@ -32,7 +32,6 @@ public class VoteController {
   @Operation(summary = "투표 생성", description = "관리자 전용, 일요일만 투표 생성 가능")
   @PostMapping
   public ResponseEntity<Void> createVote() {
-    //VoteOptionsDto voteCreateDtos = voteService.createVote();
     voteService.createVote();
     return ResponseEntity.noContent().build();
   }
@@ -46,7 +45,7 @@ public class VoteController {
   @Operation(summary = "현재 진행중인 투표에 참여하기")
   @PostMapping("/participate")
   public ResponseEntity<Void> participateVote(
-      @Valid @RequestBody VoteParticipationRequestDto voteParticipationRequestDto) {
+      @Valid @RequestBody VoteParticipationDTO voteParticipationRequestDto) {
     voteService.participateVote(voteParticipationRequestDto);
     return ResponseEntity.noContent().build();
   }
